backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
python-dotenv==1.1.1
dj-database-url==2.3.0
django-filter==24.3
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.27.2

numpy==2.3.3
pandas==2.3.2
Pillow==11.3.0
openpyxl==3.1.5

face-recognition==1.3.0
dlib==20.0.0
opencv-python==4.11.0.86
scikit-image==0.25.0

scikit-learn==1.6.1

psycopg2-binary==2.9.10
